requirements.txt
requests==2.31.0
pandas==2.2.2
pandera==0.19.3
boto3==1.34.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-cov==5.0.0
